Move CommonOptionsTrait.
This fixes a compatibility issue that will exist for all of 3.5.x so
I think fixing bake is the better option. I'll also fix up the
CommandScanner so this doesn't happen in the future.

Refs cakephp/cakephp#11326
