+++ src/Shell/Task/BakeTask.php
@@ -14,7 +14,7 @@
  */
 namespace Bake\Shell\Task;
 
-use Bake\Shell\CommonOptionsTrait;
+use Bake\Utility\CommonOptionsTrait;
 use Cake\Cache\Cache;
 use Cake\Console\Shell;
 use Cake\Core\Configure;
